Unified and improved decoding of arrays and strings
If assumed that a string in a S7-1200 has no special type-flags and is
transmitted as an array of USInt, then the dissection can be unified. If
so, the set type-flag bit 0x10 marks that an array with a size-header is
following. With this changes, the deconding has not to differ between
session and normal telegrams.